backend/python_service/requirements.txt
fastapi
uvicorn[standard]
pytesseract
pdf2image
Pillow
python-multipart
pydantic
regex
